ну тут по красоте с файлами работу почти закончила (редактирование и скачивание работает исправно)

+++ ABIS.Main/ABIS.Data/Models/File.cs
@@ -35,7 +35,7 @@ namespace ABIS.Data.Models
         /// <summary>
         /// Размер фйла в байтах
         /// </summary>
-        public int Size { get; set; }
+        public Int64 Size { get; set; }
 
         /// <summary>
         /// Тип файла
